Refine SSO Rails model boundaries
